Network abstruction for supporting multiple network types
Re-arrange the network setup codebase to allow support for multiple network
types with various binding methods.
